fix: 자동이체 cumulativeTotal 반영 누락 수정 (Processor 분리) #14
